Commuincation class updated, View model added for every action

+++ app/src/main/java/com/tripewise/utilites/storage/dao/BillDao.java
@@ -21,5 +21,5 @@ public interface BillDao {
     long insertBilldata(BillData data);
 
     @Update
-    long updateBillDataData(BillData data);
+    int updateBillDataData(BillData data);
 }
